requirements.txt
AuthorizeSauce==0.3.1
BeautifulSoup==3.2.1
cssselect==0.9.1
Django==1.6.5
MySQL-python==1.2.5
PyYAML==3.10
South==0.8.4
beautifulsoup4==4.1.3
bleach==1.4
boto==2.25.0
celery==3.1.18
coverage==3.5.3
defusedxml==0.4.1
django-celery==3.1.16
django-debug-toolbar==1.2.1
django-extensions==1.3.3
django-haystack==2.1.0
django-jenkins==0.15.0
django-passwords==0.2.0
django-memcached==0.1.2
django-nose==1.1
django-recaptcha==1.0
django-storages==1.1.4
django-taggit==0.12.1
django-tastypie==0.9.15
django-widget-tweaks==1.1.2
django-cors-headers==1.1.0
factory_boy==2.5.0
git+git://github.com/DirectEmployers/django-fsm-widget.git@dc928b0b034a3f4c456e83b3e4511c53a24b98a5
git+git://github.com/DirectEmployers/saved-search.git@1b78ebb785874956fa2805ec16f880165caae243
git+git://github.com/DirectEmployers/solrsitemap.git@d6967443a17ace8ffff550b536b3b4f133404b3a
git+git://github.com/DirectEmployers/unicode-slugify.git
git+git://github.com/arneb/django-messages.git@bc443623fde4945a841c48665b477a529f4ec676
git+git://github.com/frankban/django-endless-pagination.git@725bde91db6da2be0663652b0df4501ba72306e7
jira==0.42
lxml==3.2.4
markdown2==2.1.0
mock==1.0.1
newrelic==2.50.0.39
ordereddict==1.1
pep8==1.3.3
pydns==2.3.6
pyflakes==0.5.0
pynliner==0.5.1
pysolr==3.2.0
python-memcached==1.53
requests==2.9.1
selenium==2.46.0
unicodecsv==0.9.4
bidict==0.3.1
validate-email==1.2
testfixtures==3.0.0
django-compressor==1.4
Markdown==2.4
xlrd==0.9.3
Sphinx==1.3.3
XlsxWriter==0.8.4
django-ajax-selects==1.5.1
git+git://github.com/DirectEmployers/django-remote-forms.git@e0a221a9249737b62fe51c4ed1b70230d6464c7d
django-impersonate==0.9.2  # 0.9.2 is the last version with Django 1.6 support
pymongo==3.2.2
html5lib==0.999 #Put in place because as of 2016-07-18 bleach depends on html5lib>=0.999 and the latest version of html5lib is screwed up.
python-dateutil==2.4.2
freezegun==0.3.7
git+git://github.com/DirectEmployers/pymongo-env.git@60ccc22cfab857ee5b7288c6fc09207ae43ffdf2
boto3==1.4.0
